Added Pickpocketing as new Locations (LootSanity!)
Just as a heads-up, this is AWFUL to play with for short &/or sync games. Could maybe be ok for longer &/or async games in its current implementation, but I don't think people will want to look for guards w/ only a 20% to have loot ON TOP OF a then 5% to be the rarest loot for that guard's loot table.

That all being said, that is just me commenting on the User Experience of this feature in its current form; as for the functionality of the feature itself it works great! I am thinking about possibly/eventually changing how the new "hybrid regions" for these pickpocketing locations are connected to the "base regions" so that the first 2/6 loot for each guard is "available" with 1 progressive episode item, the 2nd 2/6 loot is open with 2 progressive episode items, and the 3rd and final set of 2/6 loot that only has that 5% chance of spawning requiring 3 progressive episode items to be needed to "unlock" these locations. I'm not quite sure about that yet, but regardless I just wanna get this commit out there first and figure out the rest later. Will list more of my ideas for possible additions to this feature in the pull request later.