Adjusted Connection Rules for Hybrid-Regions for Pickpocketing Locations
Makes it so that common loot (30% spawn rate) is represented as "accessible" with 1 Progressive Episode Item, uncommon loot (15% spawn rate) is "accessible" with 2 Progressives, and rare loot (5% spawn rate) "require" 3 Progressives to be "available."

These changes to the hybrid-regions' connections were made on a per Episode basis, so if a piece of loot is more common to spawn in one episode over another, that location will require more or less Progressive Episode Items for each of those Episodes respectively. (Example: The Medals now need a player to have either three Episode 2 or 3 items, or one Episode 4 or 5 items.)

This change was made to make it less likely that early-game, important items get locked behind the 5% drop chance, which can be even more time consuming and possibly frustrating to get when only 20-40% of guards spawn with any loot at all.